api/requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
PyYAML>=5.4.1
python-dateutil>=2.8.2
brotli>=1.0.9
pygments
markdownify
pyppeteer
aiohttp
fastapi>=0.68.0
uvicorn==0.23.2
python-multipart
pydantic>=1.8.0
